% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_wrapper.R
\name{py_add_wrapper}
\alias{py_add_wrapper}
\title{Add two numbers using Python}
\usage{
py_add_wrapper(a, b)
}
\arguments{
\item{a}{A numeric value}

\item{b}{A numeric value}
}
\value{
The sum of a and b
}
\description{
This function uses a Python backend to add two numbers.
}
\examples{
py_add_wrapper(2, 3)
}
