% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Set up Python Environment}
\usage{
setup_python_env(
  envpath = NULL,
  packages = c("numpy", "pandas"),
  python_version = NULL
)
}
\arguments{
\item{envpath}{Path where the virtual environment should be created}

\item{packages}{Character vector of Python packages to install}

\item{python_version}{The Python version to use. If NULL, uses the
system's default Python.}
}
\description{
This function sets up a Python virtual environment for the package,
installs required packages, and configures reticulate to use this environment.
}
